chore(components): refactor to remove circular deps
