+++ packages/components/src/MapWithDraggablePin/MapWithDraggablePin.tsx
@@ -5,7 +5,7 @@ import L from 'leaflet'
 import { Button, OsmGeocoding } from '../'
 import { Box, Flex, Text } from 'theme-ui'
 import customMarkerIcon from 'src/assets/icons/map-marker.png'
-import type { Result } from '../OsmGeocoding/OsmGeocoding'
+import type { Result } from '../OsmGeocoding/types'
 
 const customMarker = L.icon({
   iconUrl: customMarkerIcon,
