+++ packages/components/src/OsmGeocoding/OsmGeocodingResultsList.tsx
@@ -1,6 +1,6 @@
 /** @jsxImportSource theme-ui */
 import { Box } from 'theme-ui'
-import type { Result } from './OsmGeocoding'
+import type { Result } from './types'
 
 export interface Props {
   results: Result[]
